% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.AIcalcu.R
\name{int.AIcalcu}
\alias{int.AIcalcu}
\title{Computes the association of microbiome data with covariates}
\usage{
int.AIcalcu(data, ref, Mprefix, covsPrefix, binPredInd, contCovStd = FALSE)
}
\arguments{
\item{data}{Needs updating}

\item{ref}{Needs updating}

\item{Mprefix}{Needs updating}

\item{covsPrefix}{Needs updating}

\item{binPredInd}{Needs updating}

\item{contCovStd}{Needs updating}
}
\description{
This function calls a custom version of the native R function IFAA from the IFAA package.
}
\details{
internal function for microbiomeIFAAPooledDS and microbiomeMZILNPooled functions.
}
