% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.allUserFunc.R
\name{int.allUserFunc}
\alias{int.allUserFunc}
\title{Computes the association of microbiome data with covariates}
\usage{
int.allUserFunc()
}
\description{
This function calls a custom version of the native R function IFAA from the IFAA package.
}
\details{
internal function for microbiomeIFAAPooledDS and microbiomeMZILNPooled functions.
}
