% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.dataInfo.R
\name{int.dataInfo}
\alias{int.dataInfo}
\title{Computes the association of microbiome data with covariates}
\usage{
int.dataInfo(
  data,
  Mprefix,
  covsPrefix,
  qualifyRefTax = FALSE,
  binPredInd,
  refReadsThresh,
  SDThresh,
  SDquantilThresh,
  balanceCut
)
}
\description{
This function calls a custom version of the native R function IFAA from the IFAA package.
}
\details{
internal function for microbiomeIFAAPooledDS and microbiomeMZILNPooled functions.
}
