% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.dataRecovTrans.R
\name{int.dataRecovTrans}
\alias{int.dataRecovTrans}
\title{Computes the association of microbiome data with covariates}
\usage{
int.dataRecovTrans(
  data,
  ref,
  Mprefix,
  covsPrefix,
  binPredInd,
  contCovStd = FALSE,
  xOnly = FALSE,
  yOnly = FALSE
)
}
\description{
This function calls a custom version of the native R function IFAA from the IFAA package.
}
\details{
internal function for microbiomeIFAAPooledDS and microbiomeMZILNPooled functions.
}
