% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.dataSparsCheck.R
\name{int.dataSparsCheck}
\alias{int.dataSparsCheck}
\title{Computes the association of microbiome data with covariates}
\usage{
int.dataSparsCheck(data, Mprefix)
}
\description{
This function calls a custom version of the native R function IFAA from the IFAA package.
}
\details{
internal function for microbiomeIFAAPooledDS and microbiomeMZILNPooled functions.
}
