% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.metaData.R
\name{int.metaData}
\alias{int.metaData}
\title{Computes the association of microbiome data with covariates}
\usage{
int.metaData(
  MicrobData,
  CovData,
  linkIDname,
  sampleIDname,
  testCov = NULL,
  ctrlCov = NULL,
  testMany = TRUE,
  ctrlMany = FALSE,
  MZILN = FALSE,
  taxDropThresh,
  standardize
)
}
\arguments{
\item{MicrobData}{Needs update}

\item{CovData}{Needs update}

\item{linkIDname}{Needs update}

\item{sampleIDname}{Needs update}

\item{testCov}{Needs update}

\item{ctrlCov}{Needs update}

\item{testMany}{Needs update}

\item{ctrlMany}{Needs update}

\item{MZILN}{Needs update}

\item{taxDropThresh}{Needs update}

\item{standardize}{Needs update}
}
\description{
This function calls a custom version of the native R function IFAA from the IFAA package.
}
\details{
internal function for microbiomeIFAAPooledDS and microbiomeMZILNPooled functions.
}
