% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.Regulariz_MZILN.R
\name{int.Regulariz_MZILN}
\alias{int.Regulariz_MZILN}
\title{Needs editing}
\usage{
int.Regulariz_MZILN(
  data,
  nRef,
  sub_taxa,
  testCovInd,
  testCovInOrder,
  testCovInNewNam,
  microbName,
  refTaxa,
  adjust_method,
  paraJobs,
  binaryInd,
  binaryInd_test,
  covsPrefix,
  Mprefix,
  fdrRate,
  sequentialRun,
  allFunc = allFunc,
  phase1_taxon_num = 200,
  trans_x_col = 200,
  spar_cutoff = 10
)
}
\arguments{
\item{data}{is a string character of the SummarizedExperiment object}

\item{adjust_method}{The adjusting method for p value adjustment. Default is "BY" for dependent FDR adjustment. It can take any adjustment method for the p.adjust function in R.}

\item{paraJobs}{If 'sequentialRun' is FALSE, this specifies the number of parallel jobs that will be registered to run the algoithm. If specified as NULL, it will automatically detect the cores to decide the number of parallel jobs.}

\item{fdrRate}{The false discovery rate for identifying taxa/OTU/ASV associated with 'covariates'.}

\item{microbVar_ds}{This takes a single or vector of microbiome variable names (e.g., taxa, OTU and ASV names) of interest. Default is "all" meaning all microbiome variables will be analyzed. If a subset of microbiome variables is specified, the output will only contain the specified variables, and p-value adjustment for multiple testing will only be applied to the subset.}

\item{refTaxa_ds}{is a string character for the microbiome variable denominator (can also be a vector of microbiome variables)}

\item{allCov_ds}{is a string character of covariates to be examined along the microbiome variables (can also be a vector of covariates).}

\item{sampleIDname}{is a string character for the sample ID variable.}

\item{taxDropThresh}{The threshold of number of non-zero sequencing reads for each taxon to be dropped from the analysis. Default is 0 which means that taxon without any sequencing reads will be dropped from the analysis.}

\item{standardize}{is a logical. If TRUE, the design matrix for X will be standardized in the analyses and the results. Default is FALSE.}

\item{verbose}{Whether the process message is printed out to the console. Default is TRUE.}
}
\description{
Needs editing
}
\details{
Needs editing

internal function for the microbiomeMZILNPooled function.
}
