% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeBetaregDS.R
\name{microbiomeBetaregDS}
\alias{microbiomeBetaregDS}
\title{Fits beta regression models for rates and proportions via maximum likelihood}
\usage{
microbiomeBetaregDS(
  df,
  formula,
  na.action,
  weights,
  offset,
  link,
  link.phi,
  type,
  control,
  model,
  x,
  y
)
}
\arguments{
\item{df}{is a string character for the data.frame object}

\item{formula}{symbolic description of the model.}

\item{na.action}{is a string character specifying how NA data shall be treated.}

\item{weights}{optional numeric vector of case weights.}

\item{offset}{optional numeric vector with an a priori known component to be included in the linear predictor for the mean. In betareg.fit, offset may also be a list of two offsets for the mean and precision equation, respectively.}

\item{link}{character specification of the link function in the mean model (mu). Currently, "logit", "probit", "cloglog", "cauchit", "log", "loglog" are supported. Alternatively, an object of class "link-glm" can be supplied.}

\item{link.phi}{character specification of the link function in the precision model (phi). Currently, "identity", "log", "sqrt" are supported. The default is "log" unless formula is of type y ~ x where the default is "identity" (for backward compatibility). Alternatively, an object of class "link-glm" can be supplied.}

\item{type}{character specification of the type of estimator. Currently, maximum likelihood ("ML"), ML with bias correction ("BC"), and ML with bias reduction ("BR") are supported.}

\item{control}{a list of control arguments specified via betareg.control.}

\item{model}{is a logical. If TRUE, the model frame from the model fit will be returned.}

\item{x}{is a logical. If TRUE, the response from the model fit will be returned.}

\item{y}{is a logical. If TRUE, the model matrix from the model fit will be returned.}
}
\value{
\code{microbiomeBetaregDS} returns a beta regression model
}
\description{
This function is similar to the native R function betareg from the betareg package
}
\details{
The function calls the server-side function \code{microbiomeBetaregDS} that computes the beta regression model
from data.frame. The function uses parametrization with mean and precision parameter.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
