% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeMZILNDS.R
\name{microbiomeMZILNDS}
\alias{microbiomeMZILNDS}
\title{Computes the multivariate zero-inflated logistic normal model}
\usage{
microbiomeMZILNDS(SumExp, taxa, covariates)
}
\arguments{
\item{SumExp}{is a string character of the data.frame}

\item{taxa}{is a string character for the microbiome variable denominator (can also be a vector of microbiome variables)}

\item{covariates}{is a string character of covariates to be examined along the microbiome variables (can also be a vector of covariates).}
}
\value{
XXXXX
}
\description{
This function calls the native R function from the IFAA package
}
\details{
The function computes a model from a SummerizedExperiment object with a given set of
microbiome taxa and covariates.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
