% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizedExperimentDS.R
\name{summarizedExperimentDS}
\alias{summarizedExperimentDS}
\title{Computes the SummarizedExperiment object}
\usage{
summarizedExperimentDS(microbiomeData, covariateData)
}
\arguments{
\item{microbiomeData}{is a string character of microbiome variables of interest (can be a vector of names)}

\item{covariateData}{is a string character of covariate variables of interest (can be a vector of names)}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.summarizedExperiment} or default name \code{sumExp.newobj}
}
\description{
This function calls the native R function from Bioconductor
}
\details{
The function computes a SummarizedExperiment object from a data.frame with a given set of
microbiome and covariate data.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
