% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.dataInfo.R
\name{int.dataInfo}
\alias{int.dataInfo}
\title{Computes the association of microbiome data with covariates}
\usage{
int.dataInfo(
  data,
  Mprefix,
  covsPrefix,
  qualifyRefTax = FALSE,
  binPredInd,
  refReadsThresh,
  SDThresh,
  SDquantilThresh,
  balanceCut
)
}
\arguments{
\item{data}{Needs update}

\item{Mprefix}{Needs update}

\item{covsPrefix}{Needs update}

\item{qualifyRefTax}{Needs update}

\item{binPredInd}{Needs update}

\item{refReadsThresh}{Needs update}

\item{SDThresh}{Needs update}

\item{SDquantilThresh}{Needs update}

\item{balanceCut}{Needs update}
}
\description{
This function calls a custom version of the native R function IFAA from the IFAA package.
}
\details{
internal function for microbiomeIFAAPooledDS and microbiomeMZILNPooled functions.
}
