% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.lm_sparse.R
\name{int.lm_sparse}
\alias{int.lm_sparse}
\title{Computes the association of microbiome data with covariates}
\usage{
int.lm_sparse(x, y, intercept = FALSE, tol = 1e-07)
}
\description{
This function calls a custom version of the native R function IFAA from the IFAA package.
}
\details{
internal function for microbiomeIFAAPooledDS and microbiomeMZILNPooled functions.
}
