% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.Regulariz_IFAA.R
\name{int.Regulariz_IFAA}
\alias{int.Regulariz_IFAA}
\title{Computes the association of microbiome data with covariates}
\usage{
int.Regulariz_IFAA(
  data,
  testCovInd,
  testCovInOrder,
  testCovInNewNam,
  microbName,
  sub_taxa,
  refTaxa,
  refTaxa_newNam,
  paraJobs,
  binaryInd,
  binaryInd_test,
  covsPrefix,
  Mprefix,
  fwerRate,
  allFunc,
  adjust_method,
  sequentialRun,
  phase1_taxon_num = 200,
  trans_x_col = 200,
  spar_cutoff = 10
)
}
\description{
This function calls a custom version of the native R function IFAA from the IFAA package.
}
\details{
internal function for the microbiomeIFAAPooledDS function.
}
