% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.Regulariz_MZILN.R
\name{int.Regulariz_MZILN}
\alias{int.Regulariz_MZILN}
\title{Needs editing}
\usage{
int.Regulariz_MZILN(
  data,
  nRef,
  sub_taxa,
  testCovInd,
  testCovInOrder,
  testCovInNewNam,
  microbName,
  refTaxa,
  adjust_method,
  paraJobs,
  binaryInd,
  binaryInd_test,
  covsPrefix,
  Mprefix,
  fdrRate,
  sequentialRun,
  allFunc = allFunc,
  phase1_taxon_num = 200,
  trans_x_col = 200,
  spar_cutoff = 10
)
}
\arguments{
\item{data}{Needs update}

\item{nRef}{Needs update}

\item{sub_taxa}{Needs update}

\item{testCovInd}{Needs update}

\item{testCovInOrder}{Needs update}

\item{testCovInNewNam}{Needs update}

\item{microbName}{Needs update}

\item{refTaxa}{Needs update}

\item{adjust_method}{Needs update}

\item{paraJobs}{Needs update}

\item{binaryInd}{Needs update}

\item{binaryInd_test}{Needs update}

\item{covsPrefix}{Needs update}

\item{Mprefix}{Needs update}

\item{fdrRate}{Needs update}

\item{sequentialRun}{Needs update}

\item{allFunc}{Needs update}

\item{phase1_taxon_num}{Needs update}

\item{trans_x_col}{Needs update}

\item{spar_cutoff}{Needs update}
}
\description{
Needs editing
}
\details{
Needs editing
}
