% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeIFAAPooledDS.R
\name{microbiomeIFAAPooledDS}
\alias{microbiomeIFAAPooledDS}
\title{Computes the association of microbiome data with covariates}
\usage{
microbiomeIFAAPooledDS(
  SumExp,
  microbVar_ds,
  testCov_ds,
  ctrlCov_ds,
  sampleIDname,
  testMany,
  ctrlMany,
  refTaxa_ds,
  adjust_method,
  fdrRate,
  paraJobs,
  standardize,
  taxDropThresh,
  verbose
)
}
\arguments{
\item{SumExp}{is a string character of the SummarizedExperiment object.}

\item{microbVar_ds}{is a single microbiome name or vector of microbiome variable names (e.g., taxa, OTU and ASV names) of interest. Default is "all" meaning all microbiome variables will be analyzed. If a subset of microbiome variables is specified, the output will only contain the specified variables, and p-value adjustment for multiple testing will only be applied to the subset.}

\item{testCov_ds}{is a string character of covariates to be examined along the microbiome variables (can also be a vector of covariates).}

\item{ctrlCov_ds}{is a string character for the covariates that will be adjusted in the model (can also be a vector of confounders).}

\item{sampleIDname}{is a string character for the sample ID variable.}

\item{testMany}{is a logical. If 'TRUE' and 'covariates' are set to NULL, then all variables in the 'covariates' will be used.}

\item{ctrlMany}{is a logical. If 'TRUE' and 'confounders' are set to NULL, then all variables except the 'coviariates' will be used as confounders.}

\item{refTaxa_ds}{vector of microbiome taxa or OTU or ASV names. Will be denominator(s) to the microbVar_ds.}

\item{adjust_method}{The adjusting method for p value adjustment. Default is "BY" for dependent FDR adjustment. It can take any adjustment method for the p.adjust function in R.}

\item{fdrRate}{The false discovery rate for identifying taxa/OTU/ASV associated with 'covariates'.}

\item{paraJobs}{If 'sequentialRun' is FALSE, this specifies the number of parallel jobs that will be registered to run the algoithm. If specified as NULL, it will automatically detect the cores to decide the number of parallel jobs.}

\item{standardize}{is a logical. If 'TRUE', the design matrix for X will be standardized in the analyses and the results. Default is FALSE.}

\item{taxDropThresh}{The threshold of number of non-zero sequencing reads for each taxon to be dropped from the analysis. Default is 0 which means that taxon without any sequencing reads will be dropped from the analysis.}

\item{verbose}{Whether the process message is printed out to the console. Default is TRUE.}
}
\value{
\code{microbiomeIFAAPooledDS} returns a list consisting of intermediary results, mostly matrix crossproducts, from which estimates will be calculated on the client-side, and some additional information.
}
\description{
This function calls a custom version of the native R function IFAA from the IFAA package.
}
\details{
The function computes an association from a SummarizedExperiment object with a given set of
microbiome taxa, covariates and confounders. In contrast to the native R function, there is no phase1 for random refTaxa selection.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
